style form for mobile
