+++ src/components/ItemForm/ItemForm.scss
@@ -1,0 +1,73 @@
+@use "../../styles/partials/mixins" as *;
+@use "../../styles/partials/variables" as *;
+.item {
+  &__form {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+    &:hover {
+      background-color: none;
+    }
+  }
+}
+
+.form {
+  &__section {
+    border-top: 1px solid rgba($color-instock-black, 0.2);
+    padding: 0.8rem 1.2rem;
+    display: flex;
+    flex-direction: column;
+    gap: 1.1rem;
+  }
+  &__section-title {
+    font-weight: 600;
+    font-size: 1rem;
+  }
+  &__buttons {
+    display: flex;
+    gap: 1rem;
+    background-color: rgba($color-light-grey, 0.5);
+    padding: 1.2rem;
+  }
+  &__item {
+    display: flex;
+    flex-direction: column;
+    gap: 0.375rem;
+  }
+  &__radio-container {
+    display: flex;
+    justify-content: space-between;
+  }
+  &__radio {
+    width: calc(50% - 1rem);
+    display: flex;
+    gap: 0.4rem;
+  }
+  &__label {
+    font-size: 0.7rem;
+    font-weight: 600;
+  }
+  &__radio-label {
+    font-size: 0.7rem;
+    &--grey {
+      color: $color-slate;
+    }
+  }
+  &__radio-label ::before {
+    background-color: red;
+  }
+  &__input {
+    border: 1px solid $color-cloud;
+    border-radius: 1.25rem;
+    resize: none;
+    padding: 0.5rem 0.8rem;
+    font-size: 0.7rem;
+    &--error {
+      border-color: red;
+    }
+    &--grey {
+      color: $color-slate;
+    }
+  }
+}
