+++ src/styles/partials/_global.scss
@@ -17,7 +17,7 @@
 }
 
 .page-content {
-  margin: 0 1rem;
+  margin: 0 0.8rem;
   box-shadow: 0 0 4rem rgba(0, 0, 0, 0.1);
   @include tablet {
     margin: 0 2rem;
@@ -58,8 +58,8 @@ button {
   }
   &--secondary {
     background-color: $color-white;
-    border: 2px solid $color-light-grey;
-    color: $color-graphite;
+    border: 1px solid $color-cloud;
+    color: $color-slate;
     &:hover {
       color: $color-instock-indigo;
       border-color: $color-instock-indigo;
@@ -72,6 +72,17 @@ button {
       background-color: $color-graphite;
     }
   }
+}
+
+select {
+  //remove default dropdown arrow
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  //add custom dropdown arrow
+  background-image: url("../../assets/icons/arrow_drop_down-24px.svg");
+  background-repeat: no-repeat;
+  background-position-x: 96%;
+  background-position-y: 3px;
 }
 
 a {
